Implement overlapping local brokerage (WIP)
The hard problem here is that, in case of overlapping groups, it is unclear what constitutes a node 's own group (there may be several).

We could determine this vis-a-vis each group a node belongs to or vis-a-vis an explicitly supplied group. However, this would probably complicate the API, which we do not really want to make any more complex.