Implement basic support for layered networks
This is a way to handle overlapping groups by 'expanding' nodes that are in multiple groups into separate nodes (representing a node-layer pair). These nodes are then linked to each node the original node was linked to. If node u is expanded into (u, 0) and (u, 1), we also add a link between (u, 0) and (u, 1) to avoid accidentally counting shortest paths between (u, 0) and (u, 1).

Lots remaining to do:
* This approach can work and has the advantage of not requiring any changes to the core algorithms. However, it may not be the most efficient way; e.g. the number of edges can increase quite drastically.
* We do not really handle weighted networks at the moment. Specifically, we'd need to make sure that any link (u, 0) - (u, 1) always constitutes the shortest path.
* The API is clunky. At the moment, one needs to do::

    >>> G, groups = gefura.decouple_overlap(G, groups)
    >>> gamma = gefura.global_gefura(G, groups)
    >>> gamma = aggregate_overlap(gamma)

  Ideally, the API would not change compared to previous versions.
* Probably only global gefura works at the moment.
* Directed networks are untested.
* Normalization is not yet handled.